move eval and eval_isolated to std::rl
