+++ src/interpreter/stdlib/mod.rs
@@ -3,8 +3,6 @@
 pub mod array;
 pub mod bitwise;
 mod common;
-pub mod eval;
-pub mod eval_isolated;
 pub mod fs;
 pub mod io;
 pub mod len;
@@ -13,6 +11,7 @@ pub mod path;
 pub mod process;
 pub mod random;
 pub mod result;
+pub mod rl;
 pub mod string;
 pub mod terminal;
 pub mod time;
